L1 adapt

+++ mmdet/models/utils/transformer.py
@@ -1199,8 +1199,11 @@ class DeformableDetrTransformer(Transformer):
         lvl_pos_embed_flatten = torch.cat(lvl_pos_embed_flatten, 1)
         spatial_shapes = torch.as_tensor(
             spatial_shapes, dtype=torch.long, device=feat_flatten.device)
+
+
+
         level_start_index = torch.cat((spatial_shapes.new_zeros(
-            (1, )), spatial_shapes.prod(1).cumsum(0)[:-1]))
+            (1, )), spatial_shapes.prod(1).cumsum(0)[:-1].astype(torch.long)))
         valid_ratios = torch.stack(
             [self.get_valid_ratio(m) for m in mlvl_masks], 1)
